Car Controller & Views (#68)

+++ service/src/main/java/cz/muni/fi/service/TeamServiceImpl.java
@@ -1,7 +1,10 @@
 package cz.muni.fi.service;
 
+import cz.muni.fi.dao.CarDao;
+import cz.muni.fi.dao.DriverDao;
 import cz.muni.fi.dao.TeamDao;
 import cz.muni.fi.dto.DriverDTO;
+import cz.muni.fi.entities.Car;
 import cz.muni.fi.entities.Driver;
 import cz.muni.fi.entities.Team;
 import cz.muni.fi.service.exception.ServiceDataAccessException;
@@ -17,8 +20,15 @@ import java.util.List;
  */
 @Service
 public class TeamServiceImpl implements TeamService {
+
     @Autowired
     private TeamDao teamDao;
+
+    @Autowired
+    private DriverDao driverDao;
+
+    @Autowired
+    private CarDao carDao;
 
     @Override
     public Team findTeamById(Long id) {
@@ -34,6 +44,18 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public List<Team> findAllTeams() {
         return teamDao.listTeams();
+    }
+
+    @Override
+    public List<Car> findAllTeamCars(){
+        List<Car> cars = new ArrayList<>();
+
+        for (Team team: teamDao.listTeams()) {
+            cars.add(team.getCarOne());
+            cars.add(team.getCarTwo());
+        }
+
+        return cars;
     }
 
     @Override
@@ -63,6 +85,10 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public Team createTeam(Team team) {
         if (team == null) throw new IllegalArgumentException("Team argument cannot be null!");
+        if (team.getCarOne() == null || team.getCarTwo() == null)
+            throw new IllegalArgumentException("Team car cannot be null!");
+
+        driverUpdateUtil(team, true);
 
         try {
             return teamDao.addTeam(team);
@@ -74,6 +100,12 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public Team updateTeam(Team team) {
         if (team == null) throw new IllegalArgumentException("Team argument cannot be null!");
+        if (team.getCarOne() == null || team.getCarTwo() == null)
+            throw new IllegalArgumentException("Team car cannot be null!");
+
+        Team old_team = teamDao.findTeamById(team.getId());
+        driverUpdateUtil(old_team, false);
+        driverUpdateUtil(team, true);
 
         try {
             return teamDao.updateTeam(team);
@@ -85,11 +117,29 @@ public class TeamServiceImpl implements TeamService {
     @Override
     public void deleteTeam(Team team) {
         if (team == null) throw new IllegalArgumentException("Team argument cannot be null!");
+        if (team.getCarOne() == null || team.getCarTwo() == null)
+            throw new IllegalArgumentException("Team car cannot be null!");
+
+        driverUpdateUtil(team, false);
 
         try {
             teamDao.deleteTeam(team);
         } catch (Throwable e) {
             throw new ServiceDataAccessException("Cannot delete team :" + team.toString(), e);
         }
+    }
+
+    private void driverUpdateUtil(Team team, boolean main){
+        Driver one = team.getCarOne().getDriver();
+        Driver two = team.getCarTwo().getDriver();
+        if (main){
+            one.setAsMainDriver();
+            two.setAsMainDriver();
+        } else {
+            one.setAsTestDriver();
+            two.setAsTestDriver();
+        }
+        driverDao.updateDriver(one);
+        driverDao.updateDriver(two);
     }
 }
